feat: adicionar modal reutilizável com portal e melhorias de acessibilidade
