+++ frontend/modal-repro.html
@@ -1,0 +1,85 @@
+<!DOCTYPE html>
+<html lang="pt-br">
+<head>
+<meta charset="UTF-8">
+<title>Modal overlay repro</title>
+<style>
+  body { margin: 0; font-family: sans-serif; background: #0a1f3d; color: white; }
+  .header { background: #06122b; padding: 16px; border-bottom: 1px solid rgba(255,255,255,.1); }
+  .layout-root { min-height: 100vh; }
+
+  /* Reproduces frontend/src/index.css fadeUp keyframes + animate-fade-up */
+  @keyframes fadeUp {
+    from { opacity: 0; transform: translateY(14px); }
+    to   { opacity: 1; transform: translateY(0); }
+  }
+  .animate-fade-up { animation: fadeUp 0.4s cubic-bezier(0.16,1,0.3,1) both; }
+
+  /* Reproduces <main className="mx-auto max-w-6xl px-4 py-6 animate-fade-up"> */
+  main {
+    max-width: 1152px;
+    margin: 0 auto;
+    padding: 24px 16px;
+  }
+
+  .content-box { background: rgba(255,255,255,.05); border-radius: 12px; padding: 40px; text-align: center; }
+
+  /* Modal overlay - exactly as in Usuarios.tsx etc: fixed inset-0 z-50 bg-black/50 */
+  .overlay {
+    position: fixed;
+    inset: 0;
+    z-index: 50;
+    background: rgba(0,0,0,.5);
+    display: flex;
+    align-items: flex-start;
+    justify-content: center;
+    padding: 32px 16px;
+    overflow-y: auto;
+  }
+  .dialog {
+    width: 100%;
+    max-width: 28rem;
+    border-radius: 16px;
+    background: #011140;
+    box-shadow: 0 20px 25px -5px rgba(0,0,0,.5);
+    padding: 24px;
+  }
+</style>
+</head>
+<body>
+
+  <h2 style="text-align:center; padding:8px;">ANTES (bug) — overlay preso dentro do &lt;main animate-fade-up&gt;</h2>
+  <div class="layout-root">
+    <div class="header">Header / navegação</div>
+    <main class="animate-fade-up">
+      <div class="content-box">Conteúdo da página (Usuários)</div>
+
+      <!-- Modal renderizado normalmente, dentro do <main> que tem transform via animação -->
+      <div class="overlay">
+        <div class="dialog">
+          <h3>Novo usuário</h3>
+          <p>Overlay deveria cobrir a tela toda, mas fica preso ao &lt;main&gt;.</p>
+        </div>
+      </div>
+    </main>
+  </div>
+
+  <hr style="margin:40px 0; border-color: rgba(255,255,255,.1)">
+
+  <h2 style="text-align:center; padding:8px;">DEPOIS (corrigido) — overlay via portal para document.body</h2>
+  <div class="layout-root">
+    <div class="header">Header / navegação</div>
+    <main class="animate-fade-up">
+      <div class="content-box">Conteúdo da página (Usuários)</div>
+    </main>
+  </div>
+  <!-- Modal "portaled" para fora do <main>, direto em document.body -->
+  <div class="overlay">
+    <div class="dialog">
+      <h3>Novo usuário</h3>
+      <p>Overlay agora cobre a tela inteira, incluindo o header.</p>
+    </div>
+  </div>
+
+</body>
+</html>
