+++ frontend/src/pages/Instituicoes.tsx
@@ -1,4 +1,4 @@
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import {
   instituicoesService,
@@ -8,6 +8,7 @@ import {
 import { useToast } from '../contexts/ToastContext';
 import { Button } from '../components/ui/Button';
 import { Card } from '../components/ui/Card';
+import { Modal } from '../components/ui/Modal';
 import { sanitizarUrl } from '../utils/seguranca';
 import { AddIcon } from '../components/icons';
 
@@ -45,55 +46,6 @@ const ESTADOS_BR = [
   'MG','PA','PB','PR','PE','PI','RJ','RN','RS','RO','RR','SC',
   'SP','SE','TO',
 ];
-
-// ─── Modal ────────────────────────────────────────────────────
-function Modal({ titulo, onClose, children }: { titulo: string; onClose: () => void; children: React.ReactNode }) {
-  useEffect(() => {
-    document.body.style.overflow = 'hidden';
-    return () => { document.body.style.overflow = ''; };
-  }, []);
-
-  useEffect(() => {
-    const onKey = (e: KeyboardEvent) => { if (e.key === 'Escape') onClose(); };
-    document.addEventListener('keydown', onKey);
-    return () => document.removeEventListener('keydown', onKey);
-  }, [onClose]);
-
-  return (
-    <div
-      className="fixed inset-0 z-50 overflow-y-auto bg-black/50"
-      role="button"
-      tabIndex={-1}
-      aria-label="Fechar modal"
-      onClick={onClose}
-      onKeyDown={(e) => { if (e.key === 'Escape' || e.key === 'Enter') onClose(); }}
-    >
-      <div className="flex min-h-full items-start justify-center px-4 py-8">
-        <div
-          className="w-full max-w-xl rounded-2xl bg-[#011140] shadow-xl flex flex-col"
-          style={{ maxHeight: 'calc(100vh - 2rem)' }}
-          role="dialog"
-          aria-modal="true"
-          onClick={(e) => e.stopPropagation()}
-          onKeyDown={(e) => e.stopPropagation()}
-        >
-          <div className="flex flex-shrink-0 items-center justify-between border-b border-white/8 px-6 py-4">
-            <h2 className="text-base font-semibold text-white">{titulo}</h2>
-            <button
-              onClick={onClose}
-              className="rounded-lg p-1.5 text-white/45 hover:bg-white/6 hover:text-white/70"
-            >
-              <svg className="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-                <path strokeLinecap="round" strokeLinejoin="round" d="M6 18L18 6M6 6l12 12" />
-              </svg>
-            </button>
-          </div>
-          <div className="overflow-y-auto px-6 py-5">{children}</div>
-        </div>
-      </div>
-    </div>
-  );
-}
 
 // ─── Campo do formulário ──────────────────────────────────────
 function Campo({
@@ -546,7 +498,7 @@ export function Instituicoes() {
 
       {/* Modal — Nova instituição */}
       {modalCriar && (
-        <Modal titulo="Nova instituição" onClose={() => setModalCriar(false)}>
+        <Modal titulo="Nova instituição" onClose={() => setModalCriar(false)} tamanho="xl">
           <FormInstituicao form={form} onChange={setForm} />
           <div className="mt-6 flex justify-end gap-3 border-t border-white/8 pt-4">
             <Button variant="secondary" onClick={() => setModalCriar(false)}>
@@ -561,7 +513,7 @@ export function Instituicoes() {
 
       {/* Modal — Editar */}
       {editando && (
-        <Modal titulo={`Editar — ${editando.nome}`} onClose={() => setEditando(null)}>
+        <Modal titulo={`Editar — ${editando.nome}`} onClose={() => setEditando(null)} tamanho="xl">
           <FormInstituicao form={form} onChange={setForm} />
           <div className="mt-6 flex justify-end gap-3 border-t border-white/8 pt-4">
             <Button variant="secondary" onClick={() => setEditando(null)}>
